Initial revision of fractal viewer
Displays a set of points generated by a dynamical system using OpenGL.
The image continuously updates with new points, due to calling updateGL
whenever there is idle time avaliable.  The new points are accumulated
into an opengl FBO so they add to the existing image.

A simple tone mapping operator is used to collapse the high intensities
of many summed points down to low dynamic range for display on screen;
The screen is assumed to have sRGB response.

The image displayed here is a simple Julia set.