First try at CUDA flame implementation
This implementation is fairly naive - each thread uses curand to
generate a stream of random numbers, and uses those to choose a mapping
to perform at random.  Each thread generates a bunch of points on the
fractal before exiting.  No particular attention has been paid to
optimization, so it's probably a lot slower than what is possible.