Bugfix: curand-randomised indices occasionally bogus
curand_uniform() generates random numbers in the range (0,1] which is
surprising given nearly every other RNG library has the range [0,1).
This resulted in occasional indexing into the wrong part of the maps
array, with a few stray points appearing as artifacts.

Surprisingly, using the mod operation (curand(&gen) % nMaps) doesn't
result in a speed hit here, so just do that for simplicity.